adds the cards in plain text
I'm adding this document so that people can get a quick sense of what
kind the flash cards have.

The cards in plain text don't have all the features of the Anki Package:
there are no images, and it doesn't retain which cards should be
double-sided and which cards should only be single-sided.

The question and answer are separated with a tab.